A rough overview of the components in the context of the application domain
Before looking at the diagram, let me first explain the thoughts I had prior to
coming up with this diagram.

We are going to design an ERP. What is this all about, and where lies the
business value in it?

We are supposedly making the product for a company which makes money by sending
products back and forth. This is the core domain of the application.

The basic idea is that a product P is moved from point A to point B, and in
this process, the company makes money.

The application will allow the company to make money by both being good at
organizing the workflow, and by making an existing workflow cheaper.

Let's say a product can be delivered to point C by taking the route A -> B ->
C, or the route A -> D -> C. The ERP must be capable of deciding which one is
better (in terms of money or time), and must manage the workflow it chooses.

So the whole point is supporting the movement of products around. Around what?
Around the world.

For this reason, let's first draw our first component: the World.

In a real project, I would continue by adding a Product component. I've chosen
not to yet, because this project is an educational one in nature, and I wanted
to get across new ideas about the architecture. If I had focused on the
products, I think there would be too much noise for the type of audience this
project is aimed at, a noise which would make the process of creating the
architecture and the architecture itself harder to understand.

We will introduce a Product component later, when the reader will feel at ease
with the architecture which, at that point, he will feel familiar with.

Instead of focusing on Products, we are going to focus on Users. This is,
incidentally, also the subject of the use cases we have drawn in the previous
commit. I think that any reader can relate to the concept of User.

So let's draw that component, the User.

Architecturally speaking, it's not that much of a difference between an User
and a Product. Instead of having Products moving around, we will have (for
now), users who live in the World and take different roles while interacting
with our system.

In the previous commit, we have concluded that there would be a standard way of
passing in data from the User to the system.

In terms of DDD (Domain-Driven Design), the data that gets into the system is
packed up as a Command. Since it looks like a good name, let's make a port of
the World component called Command Hub which will be used to receive and
process Commands from the outside.

Conceptually, the Command Hub will provide a Command Interface which any other
component can implement in order to send commands to the Hub, if it needs to.

We also want to make the components decoupled and thus reusable in different
types of applications, or to be able to combine them. This is achieved by
making an event-driven architecture.

For this reason, let's give the World component an Event Hub, which will
provide an Event interface.

But what will stich these components together?

This question is so pressing, because it's so central to our architecture, that
we will have to further investigate it.